Flutter Firebase & DDD Course [10] - Navigation Based on Auth State
goal: Create a spash page and AuthBloc for determining if a user is already logged in when the app is started.

Part 2
* Map the AuthBloc events to state, and differenticate screen navigaion based on the authenticated and unauthenticated states.

Bloc review↓
・Inputs (events)
・Output (state)
・Bloc Logic (bloc)

* Since we are introducing multiple pages like the spash screen, and the signin screen we are introducing a router (using the auto_route & auto_route_generator packages)

at the app startup (within the app_widget.dart tree) we should run authCheckRequested as soon as possible (via the globally provided authBloc at the top of the widget tree).